Apps/Phone: Fix for view leak while importing contacts from SIM
CRs-fixed: 184082

Progress dialog box can be dismissed by pressing the back key and home key.
But there was no OnCancelListener and OnDismissListener registered with
this progress dialog box. Register these listener so that dialog can be
dismissed when back and home key pressed otherwise it will avoid leak the view.