Settings: Call Feature Settings crashes when Voice Privacy or TTY Mode settings are selected.
CallFeatureSettings does not initialize TTY Mode Button object in
GSM/WCDMA mode. Voice Privacy is not a GSM/WCDMA Setting. But the
GUI displays this option for GSM/WCDMA phone and does not initialize the
checkbox in the code. Remove Voice Privacy option for GSM/WCDMA and
initialize TTY Mode Button object in both CDMA and GSM/WCDMA modes.