Telephony: Fix phone crash in BT MT call disconnect, enable mute/hold in GSM.
When phone is in global mode, even if GSM/UMTS signal is fed to it, modem initially
tries for CDMA but it will camp to GSM/UMTS immediately. During this process
global OTA data gets initiazed which is only valid in CDMA.

While displaying Incall Menu options, this OTA data is checked and MUTE/HOLD
options are disabled.

Similarly while disconnecting the MT call in BT, this OTA data is checked and
some processing is done which leads to phone crash.

To fix these issues, phone type is checked for CDMA before using OTA data.

CRs-Fixed: 226689,227182