Phone: Process AT+CHLD command synchronously.
When AT+CHLD=1 command is received, phone app invokes a framework API to
disconnect the ACTIVE call and answer the WAITING call. While disconnecting
the ACTIVE call, framework does a local cleanup and notifies all
CALL_STATE_CHANGED registrants. As part of this phone app is sending an +CIEV
indication to HF before sending OK to the AT command. This is failing a BT
standard compliance test case. So processing AT+CHLD command is made
synchronous so that no un solicited event is sent to HF before sending OK.

Change-Id: Ia129c4dc5410b72ebe8cbaa1fde7749e71077dd3
CRs-Fixed: 247286