Phone: Emergency Call support in LSM.
Changes done to support switching of subscriptions during
emergency call in Limited Service Mode.
Initiate emergency call on the subscription that has service.
In case both subscriptions are out of service, initiate on
voice preferred subscription.
Handle UI appropriately if subscription is switched by CM when
the voice preferred sub is not in LSM.

CRs-Fixed: 259217

Change-Id: Ic9844d790ad71e12b342906f7d641a67b393afba