Phone:HFP: Fix for wrong behavior with AT+CHUP command.
In HFP test TCA_BV_05, there will be an active call and a ringing
call. Expectation on AT+CHUP command is to end the active call.
Without current change, the indication of active call ended,
ie., CIEV (2,0) is not being sent as callsetup value is non zero.
The original Android open source is not having the check for call
setup.

Change-Id: I419b9c895d1fc156127211474f25a2b195732f18
CRs-Fixed:269557