packages/apps/Phone : Do not init ToneGenerator if call init fails
The ToneGenerator init performs the audio initialization
which is a synchronous operation and
might take few seconds depending on the modem availability.
This stalls the InCallScreen activity till
activity launch timeout causing a black screen.
Removing the ToneGenerator init
for call init failure case as it is not required.

Change-Id: I14ec1990b2bbc748f4b7e1ca22ea7ff3383fd936
CRs-Fixed: 310194