packages/apps/Phone : Do not init ToneGenerator if call init fails
The ToneGenerator init performs the audio initialization
which is a synchronous operation and
might take few seconds depending on the modem availability.
This stalls the InCallScreen activity till
activity launch timeout causing a black screen.
Removing the ToneGenerator init
for call init failure case as it is not required.

Change-Id: If1a0722cc05cf845f8b9e82729daca652eebf35a
CRs-Fixed: 310194