Phone/BT: Updating AudioManager on SCO status once the Socket is closed.
When BT SCO disconnect and Audio Manager update are made at the same time
in audioOff, it is observed the action to update Audio Path is happening
proior to closure of the BT Channel. This is leading to glitch in some
remote headsets. The current change will make sure SCO disconnect before
Audio routing switch.

CRs-Fixed: 323743

Change-Id: Icaea742099ba74fd58f85b215952a84fa104bb62