Phone: Don't call resetInCallScreenMode
Prevent InCallScreen's call buttons from being removed
after a radio technology change.

When an emergency call is initiated in out of service, it is
possible to get a radio technology change if phone is unable
to camp on the one that is currently configured. This causes
an incorrect cleanup, which removes the in-call buttons from
the UI.
CRs-Fixed: 459654

Change-Id: Idfe67fadc717d7368750d59d8a6052f73e6daf81