Bluetooth: Handling TURN OFF intent to cancel connect operation.
During turn off process existing connections will be disconnected. But
if a connection is in progress then the connection state will not
be connected. These cases are also to be handled during turn off.
The current change verifies if the RfcommConnect thread is waiting
and interrupts the thread during turn off.

CRs-Fixed: 345065

Change-Id: Ifaaf1be19b83ece1b2e1fb1129165d6c175633cb