Revert "Allow only NT_MODE_GSM_ONLY or NT_MODE_WCDMA_PREF"
The issue is that this fix is resetting the network mode in the setting
database to NT_MODE_WCDMA_PREF (0) for all non GSM_ONLY modes. So if user
sets the mode to say Global, the network mode is getting set to WCDMA as
soon as the GsmUmtsOptions Activity is getting created. This causes anyone
who uses the Settings.Secure.PREFERRED_NETWORK_MODE directly to see value
of NT_MODE_WCDMA_PREF instead of the actual network mode that is set by
the user.

This reverts commit 5c7d3167b8372b040d6c0a695f7c0822d80b2c21.
CRs-Fixed: 349523

Change-Id: Idc3aa24789e93c0685d553eb66fd13b36437a699