Settings(DSDS): Handle config change for mcc and mnc
While activating/deactivating subscription, all the options
in the multi SIM Settings becomes non-functional.

To fix this Add configChanges as mcc and mnc in SetSubscription
so that the activity is not restarted when the config changes.

CRs-Fixed: 367649
(cherry picked from commit d7cca5fff12e6f864a1831c63672bfbf8559e292)

Change-Id: I80369b19949e3388c4a4a1bf56124635cf38e377