Phone: Ignore touch during network operators search
When network search is in progress, the progress dialog is
dismissed if user taps the screen. This in-turn cancels
manual search and switches to the previous screen.
This behavior can be avoided if the dialog property is set
to ignore user touch events.
If required BACK key can be used to cancel search and move
to previous screen.

Change-Id: I74d3b1b420358016de4bb372b6f64a0d7fa4a82a
CRs-Fixed: 367631