IMS-VT: Fix second upgrade to video issue
initVideoCall function is used to perform functions
like choosing front or back camera, turning on speaker when
call type is video. But initVideoCall functionality is not
triggered when call is in ACTIVE state and call type changes.
So when call type changes from VT->VoLTE->VT, camera is not
turned on, therefore preview is not seen. Do camera settings
for initVideoCall without checking the flag, as it needs to be
done again in a video call if call type changes.

Remove INCOMING case - updateVideoCallState as initVideoCall
is not needed for INCOMING videocall. It will be done
during ACTIVE state.

Change-Id: I48cd55d55d92e879c2c22955d76c9c4bdcbd41a0
CRs-Fixed: 465622