Phone: The contact's portrait shouldn't flash
When a network call arrives, it doesn't use a flag to tell
if it is a network call, and incorrectly retrieve a phone
number from a SIP account to compare. As a result, the two
numbers aren't same. So the photo keeps refreshing.

Add a condition to judge if the call is a network call.

Change-Id: I253fa5ba21d848feaa40fcac743547621f0427a5
CRs-Fixed: BZ1311