Telephony: Route audio to Headset based on a property.
In a video call the audio path defaults to speaker, if no headset is
connected. The audio path defaults to headset, if it is connected.
Setting the persist.radio.ims.audio.output property to 1 overwrites
this behavior by disabling the speaker and routing audio to the headset.
If the persist.radio.ims.audio.output property is not set, or set to 0
then the default audio routing strategy is used.

CRs-Fixed: 461829

Change-Id: I73a07c4dd22646d1ef534d2c5cb9ea93c1071cba