Telephony: Remove Upgrade Call dialog when the request times out
While in IMS call user can send Upgrade Call request to upgrade voice
call to video call. The request displays Upgrade Call dialog on the
other side asking the user to upgrade the call. The dialog stays on
if the user chooses not to press on YES or NO buttons.

Remove the Upgrade Call dialog after the request times out in QMI-RIL.

Change-Id: Iafa72a6890397d86203b79b856c5a6ecd70ca1dc
CRs-Fixed: 375159