Phone: Handle cursor while export contacts to SIM
StartManagingCursor method allows the activity to take
care of managing the given Cursor lifecycle based on the
activity's lifecycle. When the activity is destroyed, all
managed Cursors will be closed automatically and its
causing Exception. But Export contacts to SIM need cursor
even if activity is destroyed. To avoid this do not use
StartManagingCursor.

Change-Id: I648e6878bde896cb9c9dc458c3fe9789e638a301
CRs-Fixed: 491733