Phone: Show contact photo when a voice call is in progress
The object of video call panel is null because its layout is not loaded
at all. So when it is time to show the contact photo, it could not pass
the judgement and the photo would be set invisible.

Include the layout of video call panel in the resource file so that
when it is time to show the photo, it works well.

CRs-Fixed: 508859

Change-Id: I229f67e84254399de7887736a5eb2905feea286e