Phone: Add support for IP call prefix
Add a preference screen in call settings for user to set IP call prefix.
Add layout resource and some strings resource for this preference screen.
Put the IP call extra to the intent, it is used in the Telephony framework
to check whether this call is IP call or not.

CRs-Fixed: 516041

Change-Id: I0d8052ed083aa26af7f961f94389884f52b08464