Phone: Enable incremental search for PLMN
As per existing implementation Modem sends all the PLMN list information
in single request. once all the PLMN list is available then it is displayed
in UI.
As part of these changes Modem will send incremental PLMN list available
to telephony. When the list is available from Modem to telephony, the
PLMN list is displayed in UI.The UI gets refreshed with incremental PLMN
list whenever there is an indication from Modem.

Change-Id: I77db51e5521cd6119fd8344900f388a788ec9531