Phone: Dial out with the correct sub that smart dialer has set
It dials out directly with the sub that user has set in settings,
regardless of which sub the smart dialer widget has set.

Get the correct sub that the smart dialer widget has set first.
If user does not set the voice to "Prompt User", dial out directly
with what we have got. Otherwise, show the sub selection dialog.

CRs-Fixed: 521702

Change-Id: I50340022dbc2538bcdaf345368b478b7691df9e0