Phone: Check for radio state properly
The implementation of ITelephony.isRadioOn() method checks for
service state instead of radio state. In cases of basebandless
targets, service state is not changed and would remain in the
initialized OUT_OF_SERVICE state. This causes delay in shutdown
procedure for such targets.

Hence a method isRadioOn() is added in phone interface and is
used to get the radio state.

Change-Id: I5a60aa39170f9d2863e95109dfd54cbd29ffdf47
CRs-Fixed: 502317