Phone: Device is going to home screen while running USSD code for SUB2
When dial USSD code(such as "*188#" for china mobile) through
sim 2, it will show InCallScreen. For InCallScreen, it will
call onCreate() and then onResume(). But in onCreate(), the
mPhone setted by setPhone(mApp.phone) isn't the phone of
MSimPhoneGlobals, so when syncWithPhoneState() is called in
onResume(), the status returned by syncWithPhoneState() will
be used to judge whether to call endInCallScreenSession() or
not and dismiss the InCallScreen finally.

We should use the phone obtained by mApp.getPhone(subscription)
above which can automatically get the proper phone.

CRs-Fixed: 494871

Change-Id: Ic492262daa68fe7e5b569f7f3952bac44fc82b5e