Phone: Can't answer/reject the call when at USSD content view
When dial "*130#" and show the USSD dialog, MT a call to the device
at this time and the IncallScreen will show behind the USSD dialog.
That's because the USSD dialog has the
WindowManager.LayoutParams.TYPE_SYSTEM_DIALOG and
WindowManager.LayoutParams.FLAG_DIM_BEHIND properties. So it always
show on top of screen.

We should add the function to hide the USSD dialog when a new ringing
connection comes and show it again after disconnect this connection.

CRs-Fixed: 551995

Change-Id: I5ea315c012ecd17f8d0e5e3b36299a80d6f20fc8