Phone: Fix the memory leak due to unclosed bluetooth headset objects
In the current implementation, since every BluetoothHeadset object
holds a strong reference to mBluetoothProfileServiceListener, which
is a member of every InCallScreen instance, InCallScreen is not getting
garbage-collected as long as the BluetoothHeadset object still holds
the reference. And while bluetooth is OFF, InCallScreen is not able to
call closeProfileProxy to make BluetoothHeadset to release the strong
reference. Hence the memory leak.

This change fixes this issue by not getting the bluetooth profile proxy
while bluetooth is OFF.

Change-Id: Iadd799c124a31c8f2ff573fc438f4ca76dc4d607
CRs-Fixed: 567346