Handle Dialing State during Emergency Call Dial from APM.
When Emergency call retry timer expires and the call
is still in dialing state, Emergency CallHelper assumes
the call is connected and cleanup happens.

Do not clean up until we are out of dialing state and
wait for the call connection to progress further before
cleaning up and end the Emergency call attempt.

Change-Id: I9b9c81c597599b92f10a37ba0f93b604c25902c3
CRs-Fixed: 600481