Report actual error code for multisim PIN/PUK operations.
Add APIs that report actual error code on PIN/PUK operations
The current APIs that are used to verify the PIN and PUK only
convey whether the operation succeeded or failed.

Add 2 new APIs that report the actual error code
'PASSWORD_INCORRECT' and 'GENRERIC_FAILURE' in case of failure.

Change-Id: I590fe5ecbd4a95909df40811e121e5543bb15f42
CRs-Fixed: 655989